fs:Add support for getting ZBD status.
The state of ZBD has some peculiarities, for example ZBD does not
intentionally maintain a variable named free_space_, so we need
to compute free space via a function GetFreeSpace(), which means
that the generation of ZBDSnapshot comes with an uncertain time cost.

Also, sometimes we would want to report these temporarily
calculated values outward through metrics. It is not a good idea
to periodically calculate a variable offline in a specific function.
So we added a ReportSnapshot interface to Metrics and left it up
to the debugger to decide when to report them: the user can decide
which variables to compute online via SnapshotOptions and then
report them one by one via Metrics::ReportSnapshot().

We have added a sample related to this in metrics_example.h.

Signed-off-by: Liu Ruicheng <liuruicheng.222@bytedance.com>